Consolidate column.h into table.h
