+++ cpp/src/arrow/api.h
@@ -23,7 +23,6 @@
 #include "arrow/array.h"
 #include "arrow/buffer.h"
 #include "arrow/builder.h"
-#include "arrow/column.h"
 #include "arrow/compare.h"
 #include "arrow/loader.h"
 #include "arrow/memory_pool.h"
