+++ cpp/src/arrow/ipc/feather.cc
@@ -30,12 +30,12 @@
 
 #include "arrow/array.h"
 #include "arrow/buffer.h"
-#include "arrow/column.h"
 #include "arrow/io/file.h"
 #include "arrow/ipc/feather-internal.h"
 #include "arrow/ipc/feather_generated.h"
 #include "arrow/loader.h"
 #include "arrow/status.h"
+#include "arrow/table.h"
 #include "arrow/util/bit-util.h"
 
 namespace arrow {
