+++ cpp/src/arrow/test-common.h
@@ -29,7 +29,6 @@
 #include "arrow/array.h"
 #include "arrow/buffer.h"
 #include "arrow/builder.h"
-#include "arrow/column.h"
 #include "arrow/memory_pool.h"
 #include "arrow/table.h"
 #include "arrow/test-util.h"
